Move safe_decode_json to a Felix-independent module
So that networking-calico can reliably use it
